Move the AIO_example_mgo inside AIO_examples folder
