+++ AIO_examples/mongo/backend/api/auth.go
@@ -4,9 +4,9 @@ import (
 	"github.com/kataras/iris"
 	"gopkg.in/mgo.v2/bson"
 
-	"github.com/iris-contrib/examples/AIO_examples_mgo/basic/backend/db"
-	"github.com/iris-contrib/examples/AIO_examples_mgo/basic/backend/libs"
-	"github.com/iris-contrib/examples/AIO_examples_mgo/basic/backend/models"
+	"github.com/iris-contrib/examples/AIO_examples/mongo/backend/db"
+	"github.com/iris-contrib/examples/AIO_examples/mongo/backend/libs"
+	"github.com/iris-contrib/examples/AIO_examples/mongo/backend/models"
 )
 
 type AuthAPI struct {
