+++ AIO_examples/mongo/backend/api/custom.go
@@ -1,8 +1,8 @@
 package api
 
 import (
-	"github.com/iris-contrib/examples/AIO_examples_mgo/basic/backend/db"
-	"github.com/iris-contrib/examples/AIO_examples_mgo/basic/backend/models"
+	"github.com/iris-contrib/examples/AIO_examples/mongo/backend/db"
+	"github.com/iris-contrib/examples/AIO_examples/mongo/backend/models"
 	"github.com/kataras/iris"
 )
