+++ AIO_examples/mongo/frontend/public/fonts/icons.eot
